Moves BaseSampling class from bn_inference to sampling/base
