+++ pgmpy/inference/bn_inference.py
@@ -1,112 +1,7 @@
-from pgmpy.inference import Inference
-from pgmpy.models import BayesianNetwork
-import pandas as pd
 import numpy as np
-import networkx as nx
-import itertools
+import pandas as pd
 
-
-class BayesianModelInference(Inference):
-    """
-    Inference class specific to Bayesian Models
-    """
-
-    def __init__(self, model):
-        """
-        Class to calculate probability (pmf) values specific to Bayesian Models
-
-        Parameters
-        ----------
-        model: Bayesian Model
-            model on which inference queries will be computed
-        """
-        if not isinstance(model, BayesianNetwork):
-            raise TypeError(
-                "Model expected type: BayesianNetwork, got type: ", type(model)
-            )
-        super(BayesianModelInference, self).__init__(model)
-        self._initialize_structures()
-
-        self.topological_order = list(nx.topological_sort(model))
-
-    def pre_compute_reduce(self, variable):
-        """
-        Get probability arrays for a node as function of conditional dependencies
-
-        Internal function used for Bayesian networks, eg. in BayesianModelSampling
-        and BayesianModelProbability.
-
-        Parameters
-        ----------
-        variable: Bayesian Model Node
-            node of the Bayesian network
-
-        Returns
-        -------
-        dict: dictionary with probability array for node
-            as function of conditional dependency values
-        """
-        variable_cpd = self.model.get_cpds(variable)
-        variable_evid = variable_cpd.variables[:0:-1]
-        cached_values = {}
-
-        for state_combination in itertools.product(
-            *[range(self.cardinality[var]) for var in variable_evid]
-        ):
-            states = list(zip(variable_evid, state_combination))
-            cached_values[state_combination] = variable_cpd.reduce(
-                states, inplace=False, show_warnings=False
-            ).values
-
-        return cached_values
-
-    def pre_compute_reduce_maps(self, variable):
-        """
-        Get probability array-maps for a node as function of conditional dependencies
-
-        Internal function used for Bayesian networks, eg. in BayesianModelSampling
-        and BayesianModelProbability.
-
-        Parameters
-        ----------
-        variable: Bayesian Model Node
-            node of the Bayesian network
-
-        Returns
-        -------
-        dict: dictionary with probability array-index for node as function of conditional dependency values,
-            dictionary with mapping of probability array-index to probability array.
-        """
-        variable_cpd = self.model.get_cpds(variable)
-        variable_evid = variable_cpd.variables[:0:-1]
-
-        state_combinations = [
-            tuple(sc)
-            for sc in itertools.product(
-                *[range(self.cardinality[var]) for var in variable_evid]
-            )
-        ]
-        weights_list = np.array(
-            [
-                variable_cpd.reduce(
-                    list(zip(variable_evid, sc)), inplace=False, show_warnings=False
-                ).values
-                for sc in state_combinations
-            ]
-        )
-
-        unique_weights, weights_indices = np.unique(
-            weights_list, axis=0, return_inverse=True
-        )
-
-        # convert weights to index; make mapping of state to index
-        state_to_index = dict(zip(state_combinations, weights_indices))
-
-        # make mapping of index to weights
-        index_to_weight = dict(enumerate(unique_weights))
-
-        # return mappings of state to index, and index to weight
-        return state_to_index, index_to_weight
+from pgmpy.sampling import BayesianModelInference
 
 
 class BayesianModelProbability(BayesianModelInference):
